Add files via upload
I added another parametric formula for the attenuation curve, from Li+08. This formula has been applied to JWST data in Markov+23, deriving the best-fitting parameters for a few z/sim 7 galaxies (Markov+23) and a complete JWST-selected sample of z=2-12 galaxies (Markov in prep.). NB: there might be a problem with the LMC and SMC attenuation curve in the current "GrainsWD01" model. For instance, the 2175 angstrom bump seems to be more pronounced in the LMC than in the MW curve. Maybe check it?