Initial implementation of calculating surviving stellar mass for particle and parametric galaxies.
Enables 'stellar_fraction' parameter on Grid - this enables the grid to load in a 'stellar_fraction' if avaiable. If not, this is set to an array of unity.

Adds methods on Parametric and Particle stars to calculate summed sfzh weighted by stellar_fraction. Adds helper methods on parametric and particle Galaxy to calculate these quantities.

Note: In the particle case, this may not work with all variations of current_mass/initial_mass - the implementation is unclear to me here. It does work in the cases when initial_masses is set (e.g. with the sample_sfzh function), but may have unexpected results if both initial_masses and current_masses are set.